+++ apps/web/src/components/pages/lernpfad_prototype/types.ts
@@ -1,0 +1,9 @@
+export type ExerciseId = string
+
+export interface Exercise {
+  done: boolean
+  title: string
+  next: ExerciseId | null
+}
+
+export type ExercisesRecord = Record<ExerciseId, Exercise>
