+++ apps/web/src/components/pages/lernpfad_prototype/use-local-storage.ts
@@ -1,0 +1,32 @@
+import { useEffect, useState } from 'react'
+
+export function useLocalStorage<T>(
+  key: string,
+  initialValue: T
+): [T, (value: T) => void] {
+  const [storedValue, setStoredValue] = useState<T>(initialValue)
+
+  useEffect(() => {
+    try {
+      const item = localStorage.getItem(key)
+      if (item) {
+        setStoredValue(JSON.parse(item) as T)
+      }
+    } catch (error) {
+      // eslint-disable-next-line no-console
+      console.error(error)
+    }
+  }, [key])
+
+  const setValue = (value: T) => {
+    try {
+      setStoredValue(value)
+      localStorage.setItem(key, JSON.stringify(value))
+    } catch (error) {
+      // eslint-disable-next-line no-console
+      console.error(error)
+    }
+  }
+
+  return [storedValue, setValue]
+}
